Inventory Updates - Replication and Better Server/Client Interaction
Lots of general updates for the inventory system and items. Mostly replication and better server/client functionality

+++ Source/CoopAdventure/HealthPotion.h
@@ -18,7 +18,7 @@ public:
 
 protected:
 	UPROPERTY(EditAnywhere)
-	float HealthValue;
+	int HealthValue;
 
 
 public:
